simplification of the last fixes
