+++ api/server/controllers/agents/v1.js
@@ -13,8 +13,6 @@ const {
   collectToolResourceFileIds,
   convertOcrToContextInPlace,
   stripFileIdsFromToolResources,
-  hasSupportContact,
-  resolveAgentOwnerContact,
 } = require('@librechat/api');
 const {
   Time,
@@ -50,6 +48,7 @@ const {
   resolveConfigServers,
   userCanUseMCPServers,
 } = require('~/server/services/MCP');
+const { attachOwnerContacts } = require('~/server/services/Agents/ownerContact');
 const { getMCPServersRegistry } = require('~/config');
 const { getLogStores } = require('~/cache');
 const db = require('~/models');
@@ -61,89 +60,12 @@ const systemTools = {
 };
 
 const MAX_SEARCH_LEN = 100;
-const OWNER_PERMISSION_BITS =
-  PermissionBits.VIEW | PermissionBits.EDIT | PermissionBits.DELETE | PermissionBits.SHARE;
 const escapeRegex = (str = '') => str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
 const getSafeModelParameters = (modelParameters) => {
   const { useResponsesApi } = modelParameters ?? {};
   return typeof useResponsesApi === 'boolean' ? { useResponsesApi } : {};
 };
 const hasEditBit = (permission) => (permission & PermissionBits.EDIT) === PermissionBits.EDIT;
-
-const getFirstOwnerIdsByResource = async (agents) => {
-  const resourceIds = agents
-    .filter((agent) => !hasSupportContact(agent))
-    .map((agent) => agent?._id)
-    .filter(Boolean);
-
-  if (resourceIds.length === 0) {
-    return new Map();
-  }
-
-  try {
-    const entries = await db.aggregateAclEntries([
-      {
-        $match: {
-          resourceType: ResourceType.AGENT,
-          resourceId: { $in: resourceIds },
-          principalType: PrincipalType.USER,
-          permBits: OWNER_PERMISSION_BITS,
-        },
-      },
-      { $sort: { grantedAt: 1, createdAt: 1, _id: 1 } },
-      { $group: { _id: '$resourceId', principalId: { $first: '$principalId' } } },
-    ]);
-
-    return new Map(
-      entries
-        .map((entry) => [entry?._id?.toString(), entry?.principalId?.toString()])
-        .filter(([resourceId, ownerId]) => resourceId && ownerId),
-    );
-  } catch (error) {
-    logger.warn('[/Agents] Failed to resolve agent owner ACL entries', error);
-    return new Map();
-  }
-};
-
-const attachOwnerContacts = async (agents) => {
-  if (!Array.isArray(agents) || agents.length === 0) {
-    return agents;
-  }
-
-  const ownerIdsByResource = await getFirstOwnerIdsByResource(agents);
-  const ownerIds = [
-    ...new Set(
-      agents
-        .filter((agent) => !hasSupportContact(agent))
-        .map((agent) => ownerIdsByResource.get(agent?._id?.toString()) ?? agent?.author?.toString())
-        .filter(Boolean),
-    ),
-  ];
-  let ownersById = new Map();
-  if (ownerIds.length > 0) {
-    try {
-      const users = await db.findUsers({ _id: { $in: ownerIds } }, 'name username email');
-      ownersById = new Map(users.map((user) => [user?._id?.toString(), user]));
-    } catch (error) {
-      logger.warn('[/Agents] Failed to resolve agent owner users', error);
-    }
-  }
-
-  return agents.map((agent) => {
-    if (hasSupportContact(agent)) {
-      delete agent.owner_contact;
-      return agent;
-    }
-    const ownerId = ownerIdsByResource.get(agent?._id?.toString()) ?? agent?.author?.toString();
-    const ownerContact = resolveAgentOwnerContact(agent, ownersById.get(ownerId) ?? null);
-    if (ownerContact) {
-      agent.owner_contact = ownerContact;
-    } else {
-      delete agent.owner_contact;
-    }
-    return agent;
-  });
-};
 
 const sanitizeViewerSkillScope = (agent, accessibleSkillSet) => {
   const skillScopeEnabled = agent.skills_enabled === true;
@@ -1241,6 +1163,7 @@ const uploadAgentAvatarHandler = async (req, res) => {
     const updatedAgent = await db.updateAgent({ id: agent_id }, data, {
       updatingUserId: req.user.id,
     });
+    await attachOwnerContacts([updatedAgent]);
 
     try {
       const avatarCache = getLogStores(CacheKeys.S3_EXPIRY_INTERVAL);
@@ -1344,6 +1267,8 @@ const revertAgentVersionHandler = async (req, res) => {
     if (updatedAgent.author) {
       updatedAgent.author = updatedAgent.author.toString();
     }
+
+    await attachOwnerContacts([updatedAgent]);
 
     if (updatedAgent.author !== req.user.id) {
       delete updatedAgent.author;
